B【关注】关注/取消后奖励处理

+++ src/main/java/com/weitaomi/application/controller/TaskPoolController.java
@@ -42,7 +42,7 @@ public class TaskPoolController extends BaseController {
     @RequestMapping(value = "/updateTaskPoolDto", method = RequestMethod.POST)
     public AjaxResult updateTaskPoolDto(HttpServletRequest request,Long taskPoolId, Integer isPublishNow,
                                         @RequestParam(required = false) Integer needNumber,
-                                        @RequestParam(required = false) Integer singScore,
+                                        @RequestParam(required = false) Double singScore,
                                         @RequestParam(required = false) Integer limitDay){
         Long memberId=this.getUserId(request);
         return AjaxResult.getOK(memberTaskPoolService.updateTaskPoolDto(memberId,taskPoolId, isPublishNow, needNumber, singScore,limitDay));
