+++ src/main/java/com/weitaomi/application/model/mapper/TaskPoolMapper.java
@@ -11,7 +11,7 @@ import java.util.List;
 public interface TaskPoolMapper extends IBaseMapper<TaskPool> {
     TaskPool getTaskPoolByOfficialId(@Param("officialAccountId") Long officialAccountId, @Param("isPublishNow") Integer isPublishNow);
     TaskPool getTaskPoolByArticleId(@Param("articleId") Long articleId, @Param("isPublishNow") Integer isPublishNow);
-    int updateTaskPoolWithScore(@Param("score") Integer score, @Param("taskId") Long taskId);
+    int updateTaskPoolWithScore(@Param("score") Double score, @Param("taskId") Long taskId);
     List<TaskPoolDto> getTaskPoolArticleDto(@Param("officialAccount") Long officialAccount, @Param("rowBounds") RowBounds rowBounds);
     List<TaskPoolDto> getTaskPoolAccountDto(@Param("officialAccount") Long officialAccount, @Param("rowBounds") RowBounds rowBounds);
 }
